Rename email.py to emails.py
Not good to conflict with stdlib..
